The Completed Assignment
My code puts together a standard deck of 52 cards, shuffles them using an always unique and unpredictable method. Next, the program prompts the user if they would like to have a card drawn. This prompt continues until the user says no, or the number of cards left in the deck reaches zero.